Force proper sign extension for mfc0/mfhc0 on MIPS64.
git-svn-id: svn://svn.savannah.nongnu.org/qemu/trunk@3432 c046a42c-6fe2-441c-8c8c-71466251a162

+++ target-mips/op.c
@@ -2328,7 +2328,7 @@ void op_ctc1 (void)
 
 void op_mfc1 (void)
 {
-    T0 = WT0;
+    T0 = (int32_t)WT0;
     DEBUG_FPU_STATE();
     RETURN();
 }
@@ -2356,7 +2356,7 @@ void op_dmtc1 (void)
 
 void op_mfhc1 (void)
 {
-    T0 = WTH0;
+    T0 = (int32_t)WTH0;
     DEBUG_FPU_STATE();
     RETURN();
 }
